Logaddexp for complex in CPU
